pyproject.toml
[project]
name = "outlook-cli"
version = "0.1.0"
description = "Alavida Outlook CLI — Graph-backed mail/calendar/contacts for AI agents"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Alavida" }]
dependencies = [
    "typer>=0.15.0",
    "rich>=13.9.0",
    "msal>=1.31.0",
    "msgraph-sdk>=1.16.0",
    "azure-identity>=1.19.0",
    "python-dotenv>=1.0.0",
    "keyring>=25.5.0",
]

[project.scripts]
outlook = "outlook_cli.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/outlook_cli"]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
